refactor(CreateSection): Refactored CreateSection and moved recalibration to SectionBuilder
