+++ AdSecCore/Builders/SectionBuilder.cs
@@ -1,35 +1,31 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+
+using AdSecGH.Parameters;
 
 using Oasys.AdSec;
+using Oasys.AdSec.DesignCode;
 using Oasys.AdSec.Materials;
+using Oasys.AdSec.Reinforcement;
 using Oasys.AdSec.Reinforcement.Groups;
 using Oasys.AdSec.StandardMaterials;
 using Oasys.Profiles;
 
 using OasysUnits;
+using OasysUnits.Units;
 
 namespace AdSecCore.Builders {
 
   public class SectionBuilder : IBuilder<ISection> {
+    private readonly IConcrete defaultMaterial = Concrete.IS456.Edition_2000.M10;
 
     private readonly List<IGroup> ReinforcementGroups = new List<IGroup>();
-    private readonly IConcrete defaultMaterial = Concrete.IS456.Edition_2000.M10;
+    private IMaterial _material;
+    private IProfile _profile;
     private ISection section;
     private SectionType sectionType;
-    private IProfile _profile;
-    private IMaterial _material;
     private double _width { get; set; }
     private double _depth { get; set; }
-
-    public static ISection Get100Section() {
-      var section = new SectionBuilder().CreateRectangularSection().WithHeight(100).WithWidth(100).Build();
-      return section;
-    }
-
-    public SectionBuilder WithMaterial(IMaterial material) {
-      _material = material;
-      return this;
-    }
 
     public ISection Build() {
       var profileBuilder = new ProfileBuilder();
@@ -44,6 +40,16 @@ namespace AdSecCore.Builders {
       }
 
       return section;
+    }
+
+    public static ISection Get100Section() {
+      var section = new SectionBuilder().CreateRectangularSection().WithHeight(100).WithWidth(100).Build();
+      return section;
+    }
+
+    public SectionBuilder WithMaterial(IMaterial material) {
+      _material = material;
+      return this;
     }
 
     private IProfile GetProfile(ProfileBuilder profileBuilder) {
@@ -65,8 +71,8 @@ namespace AdSecCore.Builders {
     }
 
     /// <summary>
-    /// Creates an invalid section for testing purposes
-    /// Given a Perimeter profile with a void polygon outside the solid polygon
+    ///   Creates an invalid section for testing purposes
+    ///   Given a Perimeter profile with a void polygon outside the solid polygon
     /// </summary>
     /// <returns></returns>
     public static ISection InvalidSection(int size = 100) {
@@ -122,11 +128,91 @@ namespace AdSecCore.Builders {
       return this;
     }
 
+    public void SetProfile(IProfile profile) { _profile = profile; }
+
+    public static List<AdSecRebarGroup> CalibrateReinforcementGroupsForIPerimeterProfile(
+      List<AdSecRebarGroup> reinforcements, IDesignCode designCodeDesignCode, IProfile profileProfile,
+      IMaterial material) {
+      var adSec = IAdSec.Create(designCodeDesignCode);
+      var sectionSection = ISection.Create(profileProfile, material);
+      var flattened = adSec.Flatten(sectionSection);
+
+      string description = profileProfile.Description();
+      string[] coordinates1 = description.Remove(0, 11).Split(new[] { ") L(", }, StringSplitOptions.None);
+      double maxY1 = double.MinValue;
+      double maxZ1 = double.MinValue;
+      foreach (string c in coordinates1) {
+        string[] value = c.Split('|');
+        double y1 = double.Parse(value[0]);
+        double z1 = double.Parse(value[1].Remove(value[1].Length - 2));
+
+        if (y1 > maxY1) {
+          maxY1 = y1;
+        }
+
+        if (z1 > maxZ1) {
+          maxZ1 = z1;
+        }
+      }
+
+      string[] coordinates2 = flattened.Profile.Description().Remove(0, 11).Split(new[] {
+        ") L(",
+      }, StringSplitOptions.None);
+      double maxY2 = double.MinValue;
+      double maxZ2 = double.MinValue;
+      foreach (string c in coordinates2) {
+        string[] value = c.Split('|');
+        double y2 = double.Parse(value[0]);
+        double z2 = double.Parse(value[1].Remove(value[1].Length - 2));
+
+        if (y2 > maxY2) {
+          maxY2 = y2;
+        }
+
+        if (z2 > maxZ2) {
+          maxZ2 = z2;
+        }
+      }
+
+      double deltaY = maxY2 - maxY1;
+      double deltaZ = maxZ2 - maxZ1;
+
+      var updatedReinforcement = new List<AdSecRebarGroup>();
+      foreach (var group in reinforcements) {
+        var duplicate = new AdSecRebarGroup();
+        if (group.Cover != null) {
+          duplicate.Cover = ICover.Create(group.Cover.UniformCover);
+        }
+
+        updatedReinforcement.Add(duplicate);
+
+        switch (group.Group) {
+          case ISingleBars bars:
+            var bundle = IBarBundle.Create(bars.BarBundle.Material, bars.BarBundle.Diameter,
+              bars.BarBundle.CountPerBundle);
+            var singleBars = ISingleBars.Create(bundle);
+
+            foreach (var point in bars.Positions) {
+              var p = IPoint.Create(new Length(point.Y.As(LengthUnit.Meter) - deltaY, LengthUnit.Meter),
+                new Length(point.Z.As(LengthUnit.Meter) - deltaZ, LengthUnit.Meter));
+              singleBars.Positions.Add(p);
+            }
+
+            duplicate.Group = singleBars;
+            break;
+
+          default:
+            duplicate.Group = group.Group;
+            break;
+        }
+      }
+
+      return updatedReinforcement;
+    }
+
     internal enum SectionType {
       Square,
       Rectangular,
     }
-
-    public void SetProfile(IProfile profile) { _profile = profile; }
   }
 }
